Rename all files to fit snake_case.
