+++ lib/main.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'pages/HomePage.dart';
-import 'pages/QRReaderPage.dart';
-import 'pages/SignInPage.dart';
-import 'pages/SignUpPage.dart';
+import 'pages/home_page.dart';
+import 'pages/qr_reader_page.dart';
+import 'pages/sign_in_page.dart';
+import 'pages/sign_up_page.dart';
 import 'providers/otp_provider.dart';
-import 'pages/LoadingPage.dart';
+import 'pages/loading_page.dart';
 
 void main() {
   runApp(ChangeNotifierProvider(
